拆分redis client 类
