+++ redis-pro/Common/RedisClient/RedisClientScan.swift
@@ -6,3 +6,17 @@
 //
 
 import Foundation
+
+extension RediStackClient {
+    
+    func isMatchAll(_ keywords:String?) -> Bool {
+        guard let keywords = keywords else {
+            return true
+        }
+        return keywords.isEmpty || keywords == "*" || keywords == "**"
+    }
+    
+    func isScan(_ keywords:String) -> Bool {
+        return keywords.isEmpty || keywords.contains("*") || keywords.contains("?")
+    }
+}
